+++ db4e/App.py
@@ -37,26 +37,24 @@ from db4e.Modules.PaneMgr import PaneMgr
 from db4e.Modules.InstallMgr import InstallMgr
 from db4e.Modules.Helper import get_radio_map
 from db4e.Modules.MessageRouter import MessageRouter
+from db4e.Modules.HealthMgr import HealthMgr
 from db4e.Messages.SubmitFormData import SubmitFormData
 from db4e.Messages.UpdateTopBar import UpdateTopBar
 from db4e.Messages.RefreshNavPane import RefreshNavPane
 from db4e.Messages.NavLeafSelected import NavLeafSelected
 from db4e.Constants.Fields import (
-    COLORTERM_ENVIRON_FIELD, COMPONENT_FIELD, DB4E_FIELD, INSTANCE_FIELD, MONEROD_FIELD, 
-    P2POOL_FIELD, P2POOL_ID_FIELD, P2POOL_INSTANCE, RADIO_MAP, REMOTE_FIELD, TERM_ENVIRON_FIELD, 
-    TO_MODULE_FIELD, TO_METHOD_FIELD, XMRIG_FIELD
-)
+    COLORTERM_ENVIRON_FIELD, COMPONENT_FIELD, DB4E_FIELD, 
+    HEALTH_MSG_FIELD, INSTANCE_FIELD, MONEROD_FIELD, P2POOL_FIELD, P2POOL_ID_FIELD, 
+    P2POOL_INSTANCE, RADIO_MAP, REMOTE_FIELD, TERM_ENVIRON_FIELD, TO_MODULE_FIELD, 
+    TO_METHOD_FIELD, XMRIG_FIELD)
 from db4e.Constants.Labels import (
     DB4E_LABEL, DEPLOYMENTS_LABEL, MONEROD_SHORT_LABEL, NEW_LABEL, P2POOL_SHORT_LABEL,
-    XMRIG_SHORT_LABEL
-)
+    XMRIG_SHORT_LABEL)
 from db4e.Constants.Panes import (
     DB4E_PANE, MONEROD_REMOTE_PANE, NEW_MONEROD_TYPE_PANE, NEW_P2POOL_TYPE_PANE,
-    NEW_XMRIG_PANE, P2POOL_REMOTE_PANE, XMRIG_PANE
-)
+    NEW_XMRIG_PANE, P2POOL_REMOTE_PANE, XMRIG_PANE)
 from db4e.Constants.Defaults import (
-    APP_TITLE_DEFAULT, COLORTERM_DEFAULT, CSS_PATH_DEFAULT, TERM_DEFAULT
-)
+    APP_TITLE_DEFAULT, COLORTERM_DEFAULT, CSS_PATH_DEFAULT, TERM_DEFAULT)
 
 class Db4EApp(App):
     TITLE = APP_TITLE_DEFAULT
@@ -69,6 +67,7 @@ class Db4EApp(App):
         self.install_mgr = InstallMgr(config)
         self.pane_catalogue = PaneCatalogue()
         self.msg_router = MessageRouter(config)
+        self.health_mgr = HealthMgr(self.depl_mgr)
 
         initialized_flag = self.depl_mgr.is_initialized()
         self.pane_mgr = PaneMgr(
@@ -93,7 +92,7 @@ class Db4EApp(App):
     # NavPane selections are routed here
     def on_nav_leaf_selected(self, message: NavLeafSelected) -> None:
         category = message.parent
-        instance = message.leaf
+        instance = message.leaf[2:] # Strip off the status unicode + ' ' 
         #print(f"Db4eApp:on_nav_leaf_selected(): {category}/{instance}")
         if category == DEPLOYMENTS_LABEL and instance == DB4E_LABEL:
             db4e_data = self.depl_mgr.get_deployment(DB4E_FIELD)
@@ -105,6 +104,8 @@ class Db4EApp(App):
         elif category == MONEROD_SHORT_LABEL:
             monerod_data = self.depl_mgr.get_deployment_by_instance(
                 component=MONEROD_FIELD, instance=instance)
+            state, results = self.health_mgr.check_p2pool(instance=instance)
+            monerod_data[HEALTH_MSG_FIELD] = results
             if monerod_data[REMOTE_FIELD]:
                 self.pane_mgr.set_pane(name=MONEROD_REMOTE_PANE, data=monerod_data)
 
@@ -114,6 +115,8 @@ class Db4EApp(App):
         elif category == P2POOL_SHORT_LABEL:
             p2pool_data = self.depl_mgr.get_deployment_by_instance(
                 component=P2POOL_FIELD, instance=instance)
+            state, results = self.health_mgr.check_p2pool(instance=instance)
+            p2pool_data[HEALTH_MSG_FIELD] = results
             if p2pool_data[REMOTE_FIELD]:
                 self.pane_mgr.set_pane(name=P2POOL_REMOTE_PANE, data=p2pool_data)
 
@@ -132,6 +135,8 @@ class Db4EApp(App):
                 xmrig_data[P2POOL_INSTANCE] = p2pool_rec[INSTANCE_FIELD]
             else:
                 xmrig_data[P2POOL_INSTANCE] = ""
+            state, results = self.health_mgr.check_xmrig(instance=instance)
+            xmrig_data[HEALTH_MSG_FIELD] = results
             self.pane_mgr.set_pane(name=XMRIG_PANE, data=xmrig_data)
 
         else: