+++ db4e/Panes/P2PoolRemote.py
@@ -6,15 +6,18 @@ db4e/Panes/P2PoolRemote.py
     Copyright (c) 2024-2025 NadimGhaznavi <https://github.com/NadimGhaznavi/db4e>
     License: GPL 3.0
 """
+from rich import box
+from rich.table import Table
 
 from textual.containers import Container, Vertical, Horizontal
-from textual.widgets import Label, MarkdownViewer, Button, Input
+from textual.widgets import Label, MarkdownViewer, Button, Input, Static
 
 from db4e.Messages.SubmitFormData import SubmitFormData
 from db4e.Constants.Fields import (
     COMPONENT_FIELD, DELETE_DEPLOYMENT_FIELD, DEPLOYMENT_MGR_FIELD, FORM_DATA_FIELD, 
-    INSTANCE_FIELD, IP_ADDR_FIELD, P2POOL_FIELD, ORIG_INSTANCE_FIELD, REMOTE_FIELD, 
-    STRATUM_PORT_FIELD, TO_MODULE_FIELD, TO_METHOD_FIELD, UPDATE_DEPLOYMENT_FIELD
+    HEALTH_MSG_FIELD, INSTANCE_FIELD, IP_ADDR_FIELD, P2POOL_FIELD, ORIG_INSTANCE_FIELD, 
+    REMOTE_FIELD, STRATUM_PORT_FIELD, TO_MODULE_FIELD, TO_METHOD_FIELD, 
+    UPDATE_DEPLOYMENT_FIELD
 )
 from db4e.Constants.Labels import (
     DELETE_LABEL, INSTANCE_LABEL, IP_ADDR_LABEL, P2POOL_REMOTE_LABEL, 
@@ -26,7 +29,25 @@ class P2PoolRemote(Container):
     instance_input = Input(id="instance_input", restrict=f"[a-zA-Z0-9_\-]*", compact=True)
     ip_addr_input = Input(id="ip_addr_input", restrict=f"[a-z0-9._\-]*", compact=True)
     stratum_port_input = Input(id="stratum_port_input", restrict=f"[0-9]*", compact=True)
+    health_msgs = Static()
 
+    def build_health_status(self, results):
+        table = Table(show_header=True, header_style="bold cyan", style="bold green", box=box.SIMPLE)
+        table.add_column("Component", width=25)
+        table.add_column("Message")
+
+        for task in results:
+            print(task)
+            for category, msg_dict in task.items():
+                message = msg_dict["msg"]
+                if msg_dict["status"] == "good":
+                    table.add_row(f"✅ [green]{category}[/]", f"[green]{message}[/]")
+                elif msg_dict["status"] == "warn":
+                    table.add_row(f"⚠️  [yellow]{category}[/]", f"[yellow]{message}[/]")
+                elif msg_dict["status"] == "error":
+                    table.add_row(f"💥 [red]{category}[/]", f"[red]{message}[/]")
+        self.health_msgs.update(table)
+        
     def compose(self):
         # Remote P2Pool deployment form
         STATIC_CONTENT = f"This screen allows you to view and edit the deployment "
@@ -50,7 +71,12 @@ class P2PoolRemote(Container):
             Horizontal(
                 Button(label=UPDATE_LABEL, id="update_button"),
                 Button(label=DELETE_LABEL, id="delete_button")
-            ))
+            ),
+      
+            Vertical(
+                self.health_msgs,
+                id="health_box",
+            ))            
 
     def set_data(self, rec):
 
@@ -59,6 +85,7 @@ class P2PoolRemote(Container):
         self.instance_input.value = rec[INSTANCE_FIELD]
         self.ip_addr_input.value = rec[IP_ADDR_FIELD]
         self.stratum_port_input.value = rec[STRATUM_PORT_FIELD]
+        self.build_health_status(rec[HEALTH_MSG_FIELD])
         
     def on_button_pressed(self, event: Button.Pressed) -> None:
         button_id = event.button.id