+++ sample_apps/test_helper.py
@@ -1,0 +1,4 @@
+from db4e.Modules.Helper import is_valid_ip, resolve_hostname, is_valid_ip_or_hostname
+print(is_valid_ip("127.0.0.1"))
+#print(resolve_hostname("kermit.osoyalce.com"))
+print(resolve_hostname("192.168.0.1"))